Make DynamicReportWriter abstract
- Add JSONReportWriter
- Add CSVReportWriter
